pyproject.toml
[project]
name = "localrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit==1.29.0",
    "langchain==0.1.0",
    "langchain-community==0.0.10",
    "qdrant-client==1.7.0",
    "sentence-transformers==2.2.2",
    "ollama==0.1.6",
    "scrapy==2.11.0",
    "pillow==10.1.0",
    "lxml==5.1.0",
    "python-dotenv==1.0.0",
    "tqdm==4.66.1",
]
